+++ module/core/former/tests/inc/enum_unnamed_tests/test_syntax.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+#[derive(Debug, PartialEq, Clone)]
+pub enum TestEnum<T: Bound>
+{
+  Variant1(InnerScalar<T>),
+}